submit merge tx for each generated account
